requirements.txt
chalice~=1.31.0
pydantic~=2.6.3
pymongo~=4.6.2
python-dotenv~=1.0.1
bump2version~=1.0.1
bumpversion~=0.6.0
pymongoose~=1.1.0
boto3==1.34.0
botocore==1.34.0
dnspython==2.4.2
python-multipart==0.0.6